Refs #2870. Fixed Windows warnings.

+++ examples/PublishHelloWorld/HelloWorld.h
@@ -78,11 +78,11 @@ static bool write_HelloWorld(Session* session, ObjectId datawriter_id, StreamId 
 
     if (128 < stream_id)
     {
-        topic_buffer = prepare_best_effort_stream_for_topic(&session->output_best_effort_stream, datawriter_id, topic_size);
+        topic_buffer = prepare_best_effort_stream_for_topic(&session->output_best_effort_stream, datawriter_id, (uint16_t)topic_size);
     }
     else
     {
-        topic_buffer = prepare_reliable_stream_for_topic(&session->output_reliable_stream, datawriter_id, topic_size);
+        topic_buffer = prepare_reliable_stream_for_topic(&session->output_reliable_stream, datawriter_id, (uint16_t)topic_size);
     }
 
     if (topic_buffer != NULL)
